+++ src/components/WalletDropdown/index.tsx
@@ -2,14 +2,16 @@ import { useWeb3React } from '@web3-react/core'
 import { useState } from 'react'
 import styled from 'styled-components/macro'
 
+import { useUserHasAvailableClaim } from '../../state/claim/hooks'
 import DefaultMenu from './DefaultMenu'
-import { LanguageMenu } from './LanguageMenu'
+import LanguageMenu from './LanguageMenu'
 import { TransactionHistoryMenu } from './TransactionMenu'
 
-const WalletWrapper = styled.div<{ isAuthenticated: boolean }>`
+const WalletWrapper = styled.div<{ isAuthenticated: boolean; hasUnclaimed: boolean }>`
   border-radius: 12px;
   width: 320px;
-  height: ${({ isAuthenticated }) => (isAuthenticated ? '324px' : '200px')};
+  height: ${({ isAuthenticated, hasUnclaimed }) => (isAuthenticated ? (hasUnclaimed ? ' 376px' : '324px') : '200px')};
+  max-height: 388px;
   display: flex;
   flex-direction: column;
   font-size: 16px;
@@ -28,9 +30,10 @@ export enum MenuState {
 const WalletDropdown = () => {
   const { account } = useWeb3React()
   const [menu, setMenu] = useState<MenuState>(MenuState.DEFAULT)
+  const hasUnclaimed = useUserHasAvailableClaim(account)
 
   return (
-    <WalletWrapper isAuthenticated={!!account}>
+    <WalletWrapper isAuthenticated={!!account} hasUnclaimed={hasUnclaimed}>
       {menu === MenuState.TRANSACTIONS && <TransactionHistoryMenu onClose={() => setMenu(MenuState.DEFAULT)} />}
       {menu === MenuState.LANGUAGE && <LanguageMenu onClose={() => setMenu(MenuState.DEFAULT)} />}
       {menu === MenuState.DEFAULT && <DefaultMenu setMenu={setMenu} />}
